+++ src/Services/chatBot/MessageStrategy/Responses/Agreement.ts
@@ -11,7 +11,7 @@ import { MessagesEnum } from '../../MessagesEnum'
 import { MessageTypes } from '../../../whatsapp/constants/MessageTypes'
 
 export class Agreement extends ResponseContract {
-  public messageSupported: Array<string> = [MessageTypes.TEXT]
+  public messageSupported: Array<string> = [MessageTypes.TEXT, MessageTypes.INTERACTIVE]
   static readonly AGREEMENT = 'convenio'
 
   public async processMessage(message: WpMessage): Promise<void> {
@@ -32,9 +32,7 @@ export class Agreement extends ResponseContract {
     const comment = MessageHelper.getCommentFromAgreement(message.msg)
     if (place) {
       await this.session.setPlace(place)
-      await this.createService(place, comment).then((serviceId: string) => {
-        this.sendMessage(Messages.cancelService(serviceId))
-      })
+      await this.createService(place, comment)
     } else {
       await this.sendMessage(Messages.getSingleMessage(MessagesEnum.DEFAULT_MESSAGE))
     }