+++ src/Services/whatsapp/services/Official/OfficialClient.ts
@@ -16,7 +16,6 @@ import { ApiMessage } from './Constants/ApiMessage'
 import { Interactive } from './Constants/Interactive'
 import { ChatBotMessage } from '../../../../Types/ChatBotMessage'
 import { MessagesEnum } from '../../../../Services/chatBot/MessagesEnum'
-import { MsgTypes } from './Constants/MsgTypes'
 import QueueService from '../../../queue/QueueService'
 
 export class OfficialClient implements WPClientInterface {
@@ -81,6 +80,7 @@ export class OfficialClient implements WPClientInterface {
       name: 'Bot',
       enabled: true,
       description: 'Mensaje de prueba',
+      interactive: null,
     }
     return await this.sendMessage('573103794656', msg).then(() => {
       this.status = WpStates.CONNECTED
@@ -90,25 +90,7 @@ export class OfficialClient implements WPClientInterface {
   }
 
   private getInteractive(message: ChatBotMessage): Interactive | false {
-    let interactive: Interactive | false
-    switch (message.id) {
-      case MessagesEnum.GREETING:
-         interactive = {
-            type: MsgTypes.location,
-            body: {
-            text: message.message,
-            },
-            action: {
-            name: 'send_location',
-            }
-         } as Interactive
-        break
-      default:
-        interactive = false
-        break
-    }
-
-    return interactive
+    return message.interactive ?? false
   }
 
   async sendMessage(phoneNumber: string, message: ChatBotMessage): Promise<void> {
@@ -117,7 +99,7 @@ export class OfficialClient implements WPClientInterface {
 
   async text(phoneNumber: string, message: ChatBotMessage): Promise<void> {
     const phone = phoneNumber.replace('@c.us', '')
-    return new Promise<void>(async (resolve, reject) => {
+    return new Promise<void>(async (resolve, reject) => {      
       const interactive = this.getInteractive(message)
       let data: ApiMessage = {
         messaging_product: this.config.messagingProduct,
@@ -151,6 +133,8 @@ export class OfficialClient implements WPClientInterface {
             type: MessageTypes.TEXT,
             body: message.message,
             fromMe: true,
+            interactive: interactive ? interactive : null,
+            interactiveReply: null,
           })
           resolve()
         })