+++ src/Types/ChatBotMessage.ts
@@ -1,7 +1,10 @@
+import { Interactive } from "../Services/whatsapp/services/Official/Constants/Interactive"
+
 export type ChatBotMessage = {
   id: string
   name: string
   description: string
   message: string
   enabled: boolean
+  interactive: Interactive | null
 }