+++ src/Types/WpMessage.ts
@@ -1,5 +1,7 @@
 import { MessageTypes } from '../Services/whatsapp/constants/MessageTypes'
 import { WpLocation } from './WpLocation'
+import { InteractiveReply } from '../Services/whatsapp/services/Official/Constants/InteractiveReply'
+import { Interactive } from '../Services/whatsapp/services/Official/Constants/Interactive'
 
 export type WpMessage = {
   created_at: number
@@ -8,4 +10,6 @@ export type WpMessage = {
   msg: string
   processed: boolean
   location: WpLocation | null
+  interactiveReply: InteractiveReply | null
+  interactive: Interactive | null
 }